Rename `utililities` to the more idiomatic `utils`
