+++ indico_toolkit/results/predictions/span.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from ..utilities import get
+from ..utils import get
 
 if TYPE_CHECKING:
     from typing import Any, Final
